refactor(service): 파사드 서비스 명 변경
